+++ Homework 1/ShellEmulator/Classes/command_abc.py
